MNT: separate deprecated xarray test functionality
